requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.0.4
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
flake8==7.1.1
flake8-quotes==3.4.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
